backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.49.0
google-genai==1.10.0
tavily-python==0.5.0
supabase==2.13.0
python-multipart==0.0.20
pdfplumber==0.11.4
pymupdf==1.25.3
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
httpx==0.28.1
anyio==4.8.0
python-jose[cryptography]==3.3.0
slowapi==0.1.9
python-magic==0.4.27
